Support different versions of istio check
